More uint type fixes

+++ DCCTimerSTM32.cpp
@@ -306,7 +306,7 @@ int ADCee::read(uint8_t pin, bool fromISR) {
 #pragma GCC push_options
 #pragma GCC optimize ("-O3")
 void ADCee::scan() {
-  static uint id = 0;        // id and mask are the same thing but it is faster to 
+  static uint8_t id = 0;        // id and mask are the same thing but it is faster to 
   static uint16_t mask = 1;  // increment and shift instead to calculate mask from id
   static bool waiting = false;
